improved no_std preformance & pico synth example

+++ src/effects/chorus.rs
@@ -1,7 +1,6 @@
 use super::{Effect, EffectParam};
-use crate::{config::SAMPLE_RATE, SampleGen};
+use crate::{config::SAMPLE_RATE, tanh, SampleGen};
 use core::fmt::Display;
-use libm::tanhf;
 use strum::EnumIter;
 
 #[derive(Debug, Clone, Copy, Hash, PartialEq, Eq, PartialOrd, Ord, EnumIter)]
@@ -79,7 +78,7 @@ impl Chorus {
         // if echo == input_sample && input_sample != 0.0 {
         //     error!("[error] {}", self.i);
         // }
-        tanhf(chorus)
+        tanh(chorus)
     }
 
     /// sets speed, takes speed in seconds
